+++ heylogs-api/src/main/java/nbbrd/heylogs/spi/VersioningSupport.java
@@ -3,7 +3,6 @@ package nbbrd.heylogs.spi;
 import lombok.NonNull;
 import org.jspecify.annotations.Nullable;
 
-import java.util.Objects;
 import java.util.function.BiPredicate;
 import java.util.function.Function;
 import java.util.function.Predicate;
@@ -17,7 +16,7 @@ public final class VersioningSupport implements Versioning {
 
     private final @NonNull String moduleId;
 
-    private final @NonNull Function<@Nullable String, Predicate<@NonNull CharSequence>> validation;
+    private final @NonNull Function<@Nullable String, @Nullable Predicate<@NonNull CharSequence>> validation;
 
     @Override
     public @NonNull String getVersioningId() {
@@ -35,7 +34,7 @@ public final class VersioningSupport implements Versioning {
     }
 
     @Override
-    public @NonNull Predicate<CharSequence> getVersioningPredicate(@Nullable String arg) {
+    public @Nullable Predicate<CharSequence> getVersioningPredicateOrNull(@Nullable String arg) {
         return validation.apply(arg);
     }
 
@@ -43,26 +42,21 @@ public final class VersioningSupport implements Versioning {
         return versioning -> versioning.getVersioningId().equals(id);
     }
 
-    public static Predicate<CharSequence> noValidation() {
-        return text -> {
-            Objects.requireNonNull(text);
-            return false;
-        };
+    public static Function<@Nullable String, Predicate<@NonNull CharSequence>> withoutArg(Predicate<CharSequence> predicate) {
+        return arg -> arg == null ? predicate : null;
     }
 
-    public static <X> Function<@Nullable String, Predicate<@NonNull CharSequence>> composing(Function<String, X> factory, BiPredicate<X, CharSequence> predicate) {
-        return arg -> compose(factory, predicate, arg);
-    }
-
-    public static <X> Predicate<@NonNull CharSequence> compose(Function<String, X> factory, BiPredicate<X, CharSequence> predicate, String arg) {
-        if (arg != null) {
-            try {
-                X engine = factory.apply(arg);
-                return text -> predicate.test(engine, text);
-            } catch (Exception ignore) {
-                // Ignore exceptions and return no validation
+    public static <X> Function<@Nullable String, Predicate<@NonNull CharSequence>> compilingArg(Function<String, X> factory, BiPredicate<X, CharSequence> predicate) {
+        return arg -> {
+            if (arg != null) {
+                try {
+                    X engine = factory.apply(arg);
+                    return text -> predicate.test(engine, text);
+                } catch (Exception ignore) {
+                    // Ignore exceptions and return null
+                }
             }
-        }
-        return VersioningSupport.noValidation();
+            return null;
+        };
     }
 }